# openapi.json
{
  "openapi": "3.0.2",
  "servers": [],
  "info": {
    "title": "Accord Protocol",
    "version": "1.0.0",
    "description": "API for document generation and agreement automation"
  },
  "components": {
    "schemas": {
...
      "org.accordproject.protocol@1.0.0.Blob": {
        "title": "Blob",
        "description": "An instance of org.accordproject.protocol@1.0.0.Blob",
        "type": "object",
        "properties": {
          "$class": {
            "type": "string",
            "default": "org.accordproject.protocol@1.0.0.Blob",
            "pattern": "^org\\.accordproject\\.protocol@1\\.0\\.0\\.Blob$",
            "description": "The class identifier for org.accordproject.protocol@1.0.0.Blob"
          },
          "base64": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          }
        },
        "required": [
          "$class",
          "base64",
          "mimeType"
        ],
        "$decorators": {
          "description": [
            "Bytes and mime type for a blob of data, such as images, files, etc."
          ]
        }
      },
      "org.accordproject.protocol@1.0.0.Text": {
        "title": "Text",
        "description": "An instance of org.accordproject.protocol@1.0.0.Text",
        "type": "object",
        "properties": {
          "$class": {
            "type": "string",
            "default": "org.accordproject.protocol@1.0.0.Text",
            "pattern": "^org\\.accordproject\\.protocol@1\\.0\\.0\\.Text$",
            "description": "The class identifier for org.accordproject.protocol@1.0.0.Text"
          },
          "templateMark": {
            "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.Document"
          },
          "templateText": {
            "type": "string"
          }
        },
        "required": [
          "$class"
        ],
        "$decorators": {
          "description": [
            "The text for a template"
          ]
        }
      },
      "org.accordproject.protocol@1.0.0.DomainModel": {
        "title": "DomainModel",
        "description": "An instance of org.accordproject.protocol@1.0.0.DomainModel",
        "type": "object",
        "properties": {
          "$class": {
            "type": "string",
            "default": "org.accordproject.protocol@1.0.0.DomainModel",
            "pattern": "^org\\.accordproject\\.protocol@1\\.0\\.0\\.DomainModel$",
            "description": "The class identifier for org.accordproject.protocol@1.0.0.DomainModel"
          }
        },
        "required": [
          "$class"
        ]
      },
      "org.accordproject.protocol@1.0.0.CtoModel": {
        "title": "CtoModel",
        "description": "An instance of org.accordproject.protocol@1.0.0.CtoModel",
        "type": "object",
        "properties": {
          "$class": {
            "type": "string",
            "default": "org.accordproject.protocol@1.0.0.CtoModel",
            "pattern": "^org\\.accordproject\\.protocol@1\\.0\\.0\\.CtoModel$",
            "description": "The class identifier for org.accordproject.protocol@1.0.0.CtoModel"
          },
          "ctoFiles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "$class",
          "ctoFiles"
        ]
      },
      "org.accordproject.protocol@1.0.0.JsonModel": {
        "title": "JsonModel",
        "description": "An instance of org.accordproject.protocol@1.0.0.JsonModel",
        "type": "object",
        "properties": {
          "$class": {
            "type": "string",
            "default": "org.accordproject.protocol@1.0.0.JsonModel",
            "pattern": "^org\\.accordproject\\.protocol@1\\.0\\.0\\.JsonModel$",
            "description": "The class identifier for org.accordproject.protocol@1.0.0.JsonModel"
          },
          "model": {
            "$ref": "#/components/schemas/concerto.metamodel@0.4.0.Model"
          }
        },
        "required": [
          "$class"
        ]
      },
      "org.accordproject.protocol@1.0.0.TemplateModel": {
        "title": "TemplateModel",
        "description": "An instance of org.accordproject.protocol@1.0.0.TemplateModel",
        "type": "object",
        "properties": {
          "$class": {
            "type": "string",
            "default": "org.accordproject.protocol@1.0.0.TemplateModel",
            "pattern": "^org\\.accordproject\\.protocol@1\\.0\\.0\\.TemplateModel$",
            "description": "The class identifier for org.accordproject.protocol@1.0.0.TemplateModel"
          },
          "typeName": {
            "type": "string"
          },
          "sharedModel": {
            "type": "string",
            "description": "The identifier of an instance of org.accordproject.protocol@1.0.0.SharedModel"
          },
          "model": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.DomainModel"
              },
              {
                "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.CtoModel"
              },
              {
                "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.JsonModel"
              }
            ]
          }
        },
        "required": [
          "$class",
          "typeName"
        ],
        "$decorators": {
          "description": [
            "The concept declaration associated with a template"
          ]
        }
      },
      "org.accordproject.protocol@1.0.0.SharedModel": {
        "title": "SharedModel",
        "description": "An instance of org.accordproject.protocol@1.0.0.SharedModel",
        "type": "object",
        "properties": {
          "$class": {
            "type": "string",
            "default": "org.accordproject.protocol@1.0.0.SharedModel",
            "pattern": "^org\\.accordproject\\.protocol@1\\.0\\.0\\.SharedModel$",
            "description": "The class identifier for org.accordproject.protocol@1.0.0.SharedModel"
          },
          "modelId": {
            "type": "string",
            "description": "The instance identifier for this type"
          },
          "model": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.DomainModel"
              },
              {
                "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.CtoModel"
              },
              {
                "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.JsonModel"
              }
            ]
          }
        },
        "required": [
          "$class",
          "modelId",
          "model"
        ],
        "$decorators": {
          "description": [
            "A shared data model"
          ],
          "resource": []
        }
      },
      "org.accordproject.protocol@1.0.0.CodeType": {
        "title": "CodeType",
        "description": "An instance of org.accordproject.protocol@1.0.0.CodeType",
        "enum": [
          "ES2015",
          "WASM_BYTES",
          "TYPESCRIPT"
        ],
        "$decorators": {
          "description": [
            "The type (language) of code"
          ]
        }
      },
...
      "org.accordproject.protocol@1.0.0.Code": {
        "title": "Code",
        "description": "An instance of org.accordproject.protocol@1.0.0.Code",
        "type": "object",
        "properties": {
          "$class": {
            "type": "string",
            "default": "org.accordproject.protocol@1.0.0.Code",
            "pattern": "^org\\.accordproject\\.protocol@1\\.0\\.0\\.Code$",
            "description": "The class identifier for org.accordproject.protocol@1.0.0.Code"
          },
          "id": {
            "type": "string",
            "description": "The instance identifier for this type"
          },
          "type": {
            "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.CodeType"
          },
          "encoding": {
            "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.CodeEncodingType"
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "$class",
          "id",
          "type",
          "encoding",
          "value"
        ],
        "$decorators": {
          "description": [
            "Executable code"
          ]
        }
      },
      "org.accordproject.protocol@1.0.0.Logic": {
        "title": "Logic",
        "description": "An instance of org.accordproject.protocol@1.0.0.Logic",
        "type": "object",
        "properties": {
          "$class": {
            "type": "string",
            "default": "org.accordproject.protocol@1.0.0.Logic",
            "pattern": "^org\\.accordproject\\.protocol@1\\.0\\.0\\.Logic$",
            "description": "The class identifier for org.accordproject.protocol@1.0.0.Logic"
          },
          "stateType": {
            "type": "string"
          },
          "codes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.Code"
            }
          }
        },
        "required": [
          "$class",
          "codes"
        ],
        "$decorators": {
          "description": [
            "The functions for a template"
          ]
        }
      },
      "org.accordproject.protocol@1.0.0.TemplateMetadata": {
        "title": "TemplateMetadata",
        "description": "An instance of org.accordproject.protocol@1.0.0.TemplateMetadata",
        "type": "object",
        "properties": {
          "$class": {
            "type": "string",
            "default": "org.accordproject.protocol@1.0.0.TemplateMetadata",
            "pattern": "^org\\.accordproject\\.protocol@1\\.0\\.0\\.TemplateMetadata$",
            "description": "The class identifier for org.accordproject.protocol@1.0.0.TemplateMetadata"
          },
          "runtime": {
            "type": "string"
          },
          "template": {
            "type": "string"
          },
          "cicero": {
            "type": "string"
          }
        },
        "required": [
          "$class",
          "runtime",
          "template",
          "cicero"
        ]
      },
      "org.accordproject.protocol@1.0.0.Template": {
        "title": "Template",
        "description": "An instance of org.accordproject.protocol@1.0.0.Template",
        "type": "object",
        "properties": {
          "$class": {
            "type": "string",
            "default": "org.accordproject.protocol@1.0.0.Template",
            "pattern": "^org\\.accordproject\\.protocol@1\\.0\\.0\\.Template$",
            "description": "The class identifier for org.accordproject.protocol@1.0.0.Template"
          },
          "metadata": {
            "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.TemplateMetadata"
          },
          "name": {
            "type": "string",
            "description": "The instance identifier for this type"
          },
          "author": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "license": {
            "type": "string"
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "logo": {
            "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.Blob"
          },
          "templateModel": {
            "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.TemplateModel"
          },
          "text": {
            "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.Text"
          },
          "logic": {
            "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.Logic"
          },
          "sampleRequest": {
            "type": "string"
          }
        },
        "required": [
          "$class",
          "metadata",
          "name",
          "author",
          "version",
          "license",
          "templateModel",
          "text"
        ],
        "$decorators": {
          "resource": [],
          "description": [
            "An Accord Project template"
          ]
        }
      },
...
      "org.accordproject.protocol@1.0.0.Metadata": {
        "title": "Metadata",
        "description": "An instance of org.accordproject.protocol@1.0.0.Metadata",
        "type": "object",
        "properties": {
          "$class": {
            "type": "string",
            "default": "org.accordproject.protocol@1.0.0.Metadata",
            "pattern": "^org\\.accordproject\\.protocol@1\\.0\\.0\\.Metadata$",
            "description": "The class identifier for org.accordproject.protocol@1.0.0.Metadata"
          },
          "values": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.KeyValue"
            }
          }
        },
        "required": [
          "$class",
          "values"
        ],
        "$decorators": {
          "description": [
            "Generic metadata comprised of key/value pairs"
          ]
        }
      },
      "org.accordproject.protocol@1.0.0.Agreement": {
        "title": "Agreement",
        "description": "An instance of org.accordproject.protocol@1.0.0.Agreement",
        "type": "object",
        "properties": {
          "$class": {
            "type": "string",
            "default": "org.accordproject.protocol@1.0.0.Agreement",
            "pattern": "^org\\.accordproject\\.protocol@1\\.0\\.0\\.Agreement$",
            "description": "The class identifier for org.accordproject.protocol@1.0.0.Agreement"
          },
          "agreementId": {
            "type": "string",
            "description": "The instance identifier for this type"
          },
          "data": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "template": {
            "type": "string",
            "description": "The identifier of an instance of org.accordproject.protocol@1.0.0.Template"
          },
          "agreementParties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.AgreementParty"
            }
          },
          "signatures": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.Signature"
            }
          },
          "agreementStatus": {
            "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.AgreementStatusType"
          },
          "historyEntries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.HistoryEntry"
            }
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.Blob"
            }
          },
          "references": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "metadata": {
            "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.Metadata"
          }
        },
        "required": [
          "$class",
          "agreementId",
          "data",
          "template",
          "agreementParties",
          "signatures",
          "agreementStatus",
          "historyEntries",
          "attachments",
          "references",
          "metadata"
        ],
        "$decorators": {
          "resource": [],
          "description": [
            "An Accord Project Agreement, an instance of a template"
          ]
        }
      },
      "org.accordproject.protocol@1.0.0.AgreementParty": {
        "title": "AgreementParty",
        "description": "An instance of org.accordproject.protocol@1.0.0.AgreementParty",
        "type": "object",
        "properties": {
          "$class": {
            "type": "string",
            "default": "org.accordproject.protocol@1.0.0.AgreementParty",
            "pattern": "^org\\.accordproject\\.protocol@1\\.0\\.0\\.AgreementParty$",
            "description": "The class identifier for org.accordproject.protocol@1.0.0.AgreementParty"
          },
          "name": {
            "type": "string"
          },
          "signatory": {
            "type": "boolean"
          },
          "role": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "network": {
            "type": "string"
          },
          "address": {
            "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.Address"
          },
          "partyId": {
            "type": "string",
            "description": "The instance identifier for this type"
          }
        },
        "required": [
          "$class",
          "name",
          "signatory",
          "partyId"
        ],
        "$decorators": {
          "description": [
            "A Party to an Agreement"
          ]
        }
      },
      "org.accordproject.protocol@1.0.0.Address": {
        "title": "Address",
        "description": "An instance of org.accordproject.protocol@1.0.0.Address",
        "type": "object",
        "properties": {
          "$class": {
            "type": "string",
            "default": "org.accordproject.protocol@1.0.0.Address",
            "pattern": "^org\\.accordproject\\.protocol@1\\.0\\.0\\.Address$",
            "description": "The class identifier for org.accordproject.protocol@1.0.0.Address"
          },
          "streetRoad": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "suburbTownCity": {
            "type": "string"
          },
          "stateTerritoryRegion": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "country": {
            "type": "string"
          }
        },
        "required": [
          "$class",
          "streetRoad"
        ],
        "$decorators": {
          "description": [
            "An Address of an Agreement Party"
          ]
        }
      },
      "org.accordproject.protocol@1.0.0.HistoryEntry": {
        "title": "HistoryEntry",
        "description": "An instance of org.accordproject.protocol@1.0.0.HistoryEntry",
        "type": "object",
        "properties": {
          "$class": {
            "type": "string",
            "default": "org.accordproject.protocol@1.0.0.HistoryEntry",
            "pattern": "^org\\.accordproject\\.protocol@1\\.0\\.0\\.HistoryEntry$",
            "description": "The class identifier for org.accordproject.protocol@1.0.0.HistoryEntry"
          },
          "agreementStatus": {
            "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.AgreementStatusType"
          },
          "data": {
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.Metadata"
          }
        },
        "required": [
          "$class",
          "agreementStatus",
          "data",
          "metadata"
        ],
        "$decorators": {
          "description": [
            "A history entry for an Agreement"
          ]
        }
      },
      "org.accordproject.protocol@1.0.0.Signature": {
        "title": "Signature",
        "description": "An instance of org.accordproject.protocol@1.0.0.Signature",
        "type": "object",
        "properties": {
          "$class": {
            "type": "string",
            "default": "org.accordproject.protocol@1.0.0.Signature",
            "pattern": "^org\\.accordproject\\.protocol@1\\.0\\.0\\.Signature$",
            "description": "The class identifier for org.accordproject.protocol@1.0.0.Signature"
          },
          "signatory": {
            "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.AgreementParty"
          },
          "signedAt": {
            "format": "date-time",
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.Metadata"
          },
          "signatureImage": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.Blob"
            }
          }
        },
        "required": [
          "$class",
          "signatory",
          "metadata",
          "signatureImage"
        ],
        "$decorators": {
          "description": [
            "A signature of an Agreement Party"
          ]
        }
      },
...
      "org.accordproject.protocol@1.0.0.AgreementStatusType": {
        "title": "AgreementStatusType",
        "description": "An instance of org.accordproject.protocol@1.0.0.AgreementStatusType",
        "enum": [
          "DRAFT",
          "SIGNNG",
          "COMPLETED",
          "SUPERSEDED"
        ],
        "$decorators": {
          "description": [
            "Runtime status of an agreement"
          ]
        }
      },
...
      "org.accordproject.commonmark@0.5.0.Document": {
        "title": "Document",
        "description": "An instance of org.accordproject.commonmark@0.5.0.Document",
        "type": "object",
        "properties": {
          "$class": {
            "type": "string",
            "default": "org.accordproject.commonmark@0.5.0.Document",
            "pattern": "^org\\.accordproject\\.commonmark@0\\.5\\.0\\.Document$",
            "description": "The class identifier for org.accordproject.commonmark@0.5.0.Document"
          },
          "xmlns": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "nodes": {
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.Node"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.Root"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.Document"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.Child"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.Text"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.CodeBlock"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.Code"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.HtmlInline"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.HtmlBlock"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.Emph"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.Strong"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.BlockQuote"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.Heading"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.ThematicBreak"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.Softbreak"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.Linebreak"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.Link"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.Image"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.Paragraph"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.List"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.Item"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.Table"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.TableHead"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.TableBody"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.TableRow"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.HeaderCell"
                },
                {
                  "$ref": "#/components/schemas/org.accordproject.commonmark@0.5.0.TableCell"
                }
              ]
            }
          },
          "startLine": {
            "type": "integer"
          },
          "endLine": {
            "type": "integer"
          }
        },
        "required": [
          "$class",
          "xmlns"
        ]
      },
...
      "concerto.metamodel@0.4.0.Model": {
        "title": "Model",
        "description": "An instance of concerto.metamodel@0.4.0.Model",
        "type": "object",
        "properties": {
          "$class": {
            "type": "string",
            "default": "concerto.metamodel@0.4.0.Model",
            "pattern": "^concerto\\.metamodel@0\\.4\\.0\\.Model$",
            "description": "The class identifier for concerto.metamodel@0.4.0.Model"
          },
          "namespace": {
            "type": "string"
          },
          "sourceUri": {
            "type": "string"
          },
          "concertoVersion": {
            "type": "string"
          },
          "imports": {
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/concerto.metamodel@0.4.0.Import"
                },
                {
                  "$ref": "#/components/schemas/concerto.metamodel@0.4.0.ImportAll"
                },
                {
                  "$ref": "#/components/schemas/concerto.metamodel@0.4.0.ImportType"
                }
              ]
            }
          },
          "declarations": {
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/concerto.metamodel@0.4.0.Declaration"
                },
                {
                  "$ref": "#/components/schemas/concerto.metamodel@0.4.0.EnumDeclaration"
                },
                {
                  "$ref": "#/components/schemas/concerto.metamodel@0.4.0.ConceptDeclaration"
                },
                {
                  "$ref": "#/components/schemas/concerto.metamodel@0.4.0.AssetDeclaration"
                },
                {
                  "$ref": "#/components/schemas/concerto.metamodel@0.4.0.ParticipantDeclaration"
                },
                {
                  "$ref": "#/components/schemas/concerto.metamodel@0.4.0.TransactionDeclaration"
                },
                {
                  "$ref": "#/components/schemas/concerto.metamodel@0.4.0.EventDeclaration"
                }
              ]
            }
          }
        },
        "required": [
          "$class",
          "namespace"
        ]
      },
...
    }
  },
  "paths": {
    "/sharedmodels": {
      "summary": "Path used to manage the list of sharedmodels.",
      "description": "The REST endpoint/path used to list and create zero or more `sharedmodel` entities.  This path contains a `GET` and `POST` operation to perform the list and create tasks, respectively.",
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.SharedModel"
                  }
                }
              }
            },
            "description": "Successful response - returns an array of `sharedmodel` entities."
          }
        },
        "operationId": "listSharedmodels",
        "summary": "List All Sharedmodels",
        "description": "Gets a list of all `sharedmodel` entities.",
        "tags": [
          "sharedmodels"
        ]
      },
      "post": {
        "requestBody": {
          "description": "A new `sharedmodel` to be created.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.SharedModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful response."
          }
        },
        "operationId": "createSharedmodel",
        "summary": "Create a Sharedmodel",
        "description": "Creates a new instance of a `sharedmodel`.",
        "tags": [
          "sharedmodels"
        ]
      }
    },
...
    "/templates": {
      "summary": "Path used to manage the list of templates.",
      "description": "The REST endpoint/path used to list and create zero or more `template` entities.  This path contains a `GET` and `POST` operation to perform the list and create tasks, respectively.",
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.Template"
                  }
                }
              }
            },
            "description": "Successful response - returns an array of `template` entities."
          }
        },
        "operationId": "listTemplates",
        "summary": "List All Templates",
        "description": "Gets a list of all `template` entities.",
        "tags": [
          "templates"
        ]
      },
      "post": {
        "requestBody": {
          "description": "A new `template` to be created.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.Template"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful response."
          }
        },
        "operationId": "createTemplate",
        "summary": "Create a Template",
        "description": "Creates a new instance of a `template`.",
        "tags": [
          "templates"
        ]
      }
    },
...
    "/agreements": {
      "summary": "Path used to manage the list of agreements.",
      "description": "The REST endpoint/path used to list and create zero or more `agreement` entities.  This path contains a `GET` and `POST` operation to perform the list and create tasks, respectively.",
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.Agreement"
                  }
                }
              }
            },
            "description": "Successful response - returns an array of `agreement` entities."
          }
        },
        "operationId": "listAgreements",
        "summary": "List All Agreements",
        "description": "Gets a list of all `agreement` entities.",
        "tags": [
          "agreements"
        ]
      },
      "post": {
        "requestBody": {
          "description": "A new `agreement` to be created.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/org.accordproject.protocol@1.0.0.Agreement"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful response."
          }
        },
        "operationId": "createAgreement",
        "summary": "Create a Agreement",
        "description": "Creates a new instance of a `agreement`.",
        "tags": [
          "agreements"
        ]
      }
    },
...
  }
}